Bugfix (#31)
* Add if_exists option to drops, fix bind_tables bug

The first to prevent needing to do a try catch if you're not sure if the
table exists, since this way is way cheaper. The bug with
DatabaseInterface.bind_tables was that if you called it before you were
connected to the database, it would throw an exception, since it tried
to get some information about the dialect. Now we just make something
reasonable up if the database isn't connected yet.

* Update changelog

* Actually fix the error this time, nerd

* Correct changelog version

* Update docstring, make arg order agree

* Correctly skip auto-generated index names

* Ensure index names are unique across tables

* Update schema to reflect changes

* Require a DSN on init

* Remove fake index name generation

Since we can always know the dialect of our underlying database
connection now, we never need to just make up garbage.

* Update test configuration

* Don't allow a new DSN in connect

* Unbreak things the last push broke

The change to how/when the connector and dialect were saved made the
previous way to check whether the interface was connected or not stop
working. This fixes that.

* Correct links and add punctuation

* Re-add doc comments, remove partial

Instead of using functools.partial to make a function that returns a
connector instance, just store the parsed dsn and connector class so we
can instantiate it when we try to connect.

* Standardize index names

* Remove nonexistent param description

* Try to word param description better